add encript/decrypt functions
